requirements.txt
Flask==1.1.2
line-bot-sdk==1.19.0
requests==2.25.1
gunicorn==20.1.0
